feat: nuxt module
